requirements.txt
click==8.1.7
dnspython==2.7.0
markdown-it-py==3.0.0
mdurl==0.1.2
mysql-connector-python==9.1.0
psycopg2==2.9.10
Pygments==2.18.0
pymongo==4.10.1
rich==13.9.4
shellingham==1.5.4
typer==0.15.1
typing_extensions==4.12.2
azure-core==1.32.0
azure-storage-blob==12.24.0
boto3==1.35.76
botocore==1.35.76
cachetools==5.5.0
certifi==2024.8.30
cffi==1.17.1
charset-normalizer==3.4.0
click==8.1.7
cryptography==44.0.0
dnspython==2.7.0
google-api-core==2.23.0
google-auth==2.36.0
google-cloud-core==2.4.1
google-cloud-storage==2.19.0
google-crc32c==1.6.0
google-resumable-media==2.7.2
googleapis-common-protos==1.66.0
idna==3.10
isodate==0.7.2
jmespath==1.0.1
markdown-it-py==3.0.0
mdurl==0.1.2
mysql-connector-python==9.1.0
proto-plus==1.25.0
protobuf==5.29.1
psycopg2==2.9.10
pyasn1==0.6.1
pyasn1_modules==0.4.1
pycparser==2.22
Pygments==2.18.0
pymongo==4.10.1
python-dateutil==2.9.0.post0
requests==2.32.3
rich==13.9.4
rsa==4.9
s3transfer==0.10.4
shellingham==1.5.4
six==1.17.0
typer==0.15.1
typing_extensions==4.12.2
urllib3==1.26.20
